pyproject.toml
[project]
name = "lighter-cpty"
version = "0.1.0"
description = "Lighter CPTY integration for Architect"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "architect-py==5.1.3",
    "lighter-python @ git+https://github.com/elliottech/lighter-python.git",
    "grpcio==1.69.0",
    "grpcio-tools==1.69.0",
    "aiohttp==3.11.10",
    "websockets==15.0.1",
    "python-dotenv==1.0.1",
    "msgspec==0.19.0",
    "pandas==2.2.3",
    "pycryptodome==3.21.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
]